pyproject.toml
[tool.poetry]
name = "hotdog-bot"
version = "0.1.0"
description = ""
authors = ["Artem Klyuev <klprogramm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.1.0"
torchvision = "^0.16.0"
pillow = "^10.1.0"
aiogram = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
